// g6-realestate-frontend/src/app/interfaces/vivienda-detalle.interface.ts
    export interface ViviendaDetalleResponse {
        titulo: string;
        descripcion: string;
        avatar: string;
        latLng: string;
        direccion: string;
        codPostal: string;
        poblacion: string;
        provincia: string;
        tienePiscina: boolean;
        tieneAscensor: boolean;
        tieneGaraje: boolean;
        numHabitaciones: number;
        numBanos: number;
        precio: number;
        metrosCuadrados: number;
        tipo: string;
        id: number;
        nombre: string;
        apellidos: string;
        direccionPropietario: string;
        email: string;
        telefono: string;
        avatarPropietario: string;
        inmobiliariaNombre:string;
    }